submission/requirements.txt
# Core dependencies
pandas>=2.0.0
numpy>=1.24.0
pyarrow>=12.0.0
rank-bm25>=0.2.2
tqdm>=4.65.0

# Claude CLI wrapper dependencies  
requests>=2.31.0
python-dotenv>=1.0.0

# Data processing
scikit-learn>=1.3.0
scipy>=1.11.0

# API clients
arxiv>=2.0.0
semanticscholar>=0.7.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Utils
click>=8.1.0
pyyaml>=6.0
python-dateutil>=2.8.0

# Web API
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.4.0
python-multipart>=0.0.6
websockets>=12.0
slowapi>=0.1.9
aiofiles>=23.2.0

# Jupyter notebooks
jupyter>=1.0.0
jupyterlab>=4.0.0
ipywidgets>=8.1.0
nbformat>=5.9.0